backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai>=1.60.0
httpx>=0.25.0
yt-dlp
pydantic==2.5.3
